+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Development
 
 - Migrate explorer to streamlit
 - Explorer: Disable higher than daily resolutions for hosted version
+- UI: Add warming stripes
 
 0.79.0 (21.03.2024)
 *******************